Fixed some stuff

+++ js/easyui/src/jquery.form.js
@@ -1,94 +1,105 @@
 /**
+ * jQuery EasyUI 1.4.3
+ * 
+ * Copyright (c) 2009-2015 www.jeasyui.com. All rights reserved.
+ *
+ * Licensed under the GPL license: http://www.gnu.org/licenses/gpl.txt
+ * To use it on other terms please contact us at info@jeasyui.com
+ *
+ */
+/**
  * form - jQuery EasyUI
  * 
- * Copyright (c) 2009-2013 www.jeasyui.com. All rights reserved.
- *
- * Licensed under the GPL or commercial licenses
- * To use it on other terms please contact us: jeasyui@gmail.com
- * http://www.gnu.org/licenses/gpl.txt
- * http://www.jeasyui.com/license_commercial.php
  */
 (function($){
 	/**
 	 * submit the form
 	 */
 	function ajaxSubmit(target, options){
-		options = options || {};
+		var opts = $.data(target, 'form').options;
+		$.extend(opts, options||{});
 		
-		var param = {};
-		if (options.onSubmit){
-			if (options.onSubmit.call(target, param) == false) {
-				return;
-			}
-		}
+		var param = $.extend({}, opts.queryParams);
+		if (opts.onSubmit.call(target, param) == false){return;}
+		$(target).find('.textbox-text:focus').blur();
 		
-		var form = $(target);
-		if (options.url){
-			form.attr('action', options.url);
-		}
 		var frameId = 'easyui_frame_' + (new Date().getTime());
-		var frame = $('<iframe id='+frameId+' name='+frameId+'></iframe>')
-				.attr('src', window.ActiveXObject ? 'javascript:false' : 'about:blank')
-				.css({
-					position:'absolute',
-					top:-1000,
-					left:-1000
-				});
-		var t = form.attr('target'), a = form.attr('action');
-		form.attr('target', frameId);
+		var frame = $('<iframe id='+frameId+' name='+frameId+'></iframe>').appendTo('body')
+		frame.attr('src', window.ActiveXObject ? 'javascript:false' : 'about:blank');
+		frame.css({
+			position:'absolute',
+			top:-1000,
+			left:-1000
+		});
+		frame.bind('load', cb);
 		
-		var paramFields = $();
-		try {
-			frame.appendTo('body');
-			frame.bind('load', cb);
-			for(var n in param){
-				var f = $('<input type="hidden" name="' + n + '">').val(param[n]).appendTo(form);
-				paramFields = paramFields.add(f);
+		submit(param);
+		
+		function submit(param){
+			var form = $(target);
+			if (opts.url){
+				form.attr('action', opts.url);
 			}
-			form[0].submit();
-		} finally {
-			form.attr('action', a);
-			t ? form.attr('target', t) : form.removeAttr('target');
-			paramFields.remove();
+			var t = form.attr('target'), a = form.attr('action');
+			form.attr('target', frameId);
+			var paramFields = $();
+			try {
+				for(var n in param){
+					var field = $('<input type="hidden" name="' + n + '">').val(param[n]).appendTo(form);
+					paramFields = paramFields.add(field);
+				}
+				checkState();
+				form[0].submit();
+			} finally {
+				form.attr('action', a);
+				t ? form.attr('target', t) : form.removeAttr('target');
+				paramFields.remove();
+			}
+		}
+		
+		function checkState(){
+			var f = $('#'+frameId);
+			if (!f.length){return}
+			try{
+				var s = f.contents()[0].readyState;
+				if (s && s.toLowerCase() == 'uninitialized'){
+					setTimeout(checkState, 100);
+				}
+			} catch(e){
+				cb();
+			}
 		}
 		
 		var checkCount = 10;
 		function cb(){
-			frame.unbind();
-			var body = $('#'+frameId).contents().find('body');
-			var data = body.html();
-			if (data == ''){
-				if (--checkCount){
-					setTimeout(cb, 100);
-					return;
+			var f = $('#'+frameId);
+			if (!f.length){return}
+			f.unbind();
+			var data = '';
+			try{
+				var body = f.contents().find('body');
+				data = body.html();
+				if (data == ''){
+					if (--checkCount){
+						setTimeout(cb, 100);
+						return;
+					}
 				}
-				return;
-			}
-			var ta = body.find('>textarea');
-			if (ta.length){
-				data = ta.val();
-			} else {
-				var pre = body.find('>pre');
-				if (pre.length){
-					data = pre.html();
+				var ta = body.find('>textarea');
+				if (ta.length){
+					data = ta.val();
+				} else {
+					var pre = body.find('>pre');
+					if (pre.length){
+						data = pre.html();
+					}
 				}
+			} catch(e){
 			}
-			if (options.success){
-				options.success(data);
-			}
-//			try{
-//				eval('data='+data);
-//				if (options.success){
-//					options.success(data);
-//				}
-//			} catch(e) {
-//				if (options.failure){
-//					options.failure(data);
-//				}
-//			}
+			opts.success(data);
 			setTimeout(function(){
-				frame.unbind();
-				frame.remove();
+				f.unbind();
+				f.remove();
 			}, 100);
 		}
 	}
@@ -99,11 +110,6 @@
 	 * otherwise load from local data object. 
 	 */
 	function load(target, data){
-		if (!$.data(target, 'form')){
-			$.data(target, 'form', {
-				options: $.extend({}, $.fn.form.defaults)
-			});
-		}
 		var opts = $.data(target, 'form').options;
 		
 		if (typeof data == 'string'){
@@ -129,56 +135,69 @@
 			var form = $(target);
 			for(var name in data){
 				var val = data[name];
-				var rr = _checkField(name, val);
-				if (!rr.length){
-					var f = form.find('input[numberboxName="'+name+'"]');
-					if (f.length){
-						f.numberbox('setValue', val);	// set numberbox value
-					} else {
-						$('input[name="'+name+'"]', form).val(val);
-						$('textarea[name="'+name+'"]', form).val(val);
-						$('select[name="'+name+'"]', form).val(val);
+				if (!_checkField(name, val)){
+					if (!_loadBox(name, val)){
+						form.find('input[name="'+name+'"]').val(val);
+						form.find('textarea[name="'+name+'"]').val(val);
+						form.find('select[name="'+name+'"]').val(val);
 					}
 				}
-				_loadCombo(name, val);
 			}
 			opts.onLoadSuccess.call(target, data);
-			validate(target);
+			form.form('validate');
 		}
 		
 		/**
 		 * check the checkbox and radio fields
 		 */
 		function _checkField(name, val){
-			var form = $(target);
-			var rr = $('input[name="'+name+'"][type=radio], input[name="'+name+'"][type=checkbox]', form);
-			$.fn.prop ? rr.prop('checked',false) : rr.attr('checked',false);
-			rr.each(function(){
-				var f = $(this);
-				if (f.val() == String(val)){
-					$.fn.prop ? f.prop('checked',true) : f.attr('checked',true);
-				}
-			});
-			return rr;
+			var cc = $(target).find('[switchbuttonName="'+name+'"]');
+			if (cc.length){
+				cc.switchbutton('uncheck');
+				cc.each(function(){
+					if (_isChecked($(this).switchbutton('options').value, val)){
+						$(this).switchbutton('check');
+					}
+				});
+				return true;
+			}
+			cc = $(target).find('input[name="'+name+'"][type=radio], input[name="'+name+'"][type=checkbox]');
+			if (cc.length){
+				cc._propAttr('checked', false);
+				cc.each(function(){
+					if (_isChecked($(this).val(), val)){
+						$(this)._propAttr('checked', true);
+					}
+				});
+				return true;
+			}
+			return false;
+		}
+		function _isChecked(v, val){
+			if (v == String(val) || $.inArray(v, $.isArray(val)?val:[val]) >= 0){
+				return true;
+			} else {
+				return false;
+			}
 		}
 		
-		function _loadCombo(name, val){
-			var form = $(target);
-			var cc = ['combobox','combotree','combogrid','datetimebox','datebox','combo'];
-			var c = form.find('[comboName="' + name + '"]');
-			if (c.length){
-				for(var i=0; i<cc.length; i++){
-					var type = cc[i];
-					if (c.hasClass(type+'-f')){
-						if (c[type]('options').multiple){
-							c[type]('setValues', val);
+		function _loadBox(name, val){
+			var field = $(target).find('[textboxName="'+name+'"],[sliderName="'+name+'"]');
+			if (field.length){
+				for(var i=0; i<opts.fieldTypes.length; i++){
+					var type = opts.fieldTypes[i];
+					var state = field.data(type);
+					if (state){
+						if (state.options.multiple || state.options.range){
+							field[type]('setValues', val);
 						} else {
-							c[type]('setValue', val);
+							field[type]('setValue', val);
 						}
-						return;
+						return true;
 					}
 				}
 			}
+			return false;
 		}
 	}
 	
@@ -192,8 +211,16 @@
 				this.value = '';
 			} else if (t == 'file'){
 				var file = $(this);
-				file.after(file.clone().val(''));
-				file.remove();
+				if (!file.hasClass('textbox-value')){
+					var newfile = file.clone().val('');
+					newfile.insertAfter(file);
+					if (file.data('validatebox')){
+						file.validatebox('destroy');
+						newfile.validatebox();
+					} else {
+						file.remove();
+					}
+				}
 			} else if (t == 'checkbox' || t == 'radio'){
 				this.checked = false;
 			} else if (tag == 'select'){
@@ -201,25 +228,31 @@
 			}
 			
 		});
-		if ($.fn.combo) $('.combo-f', target).combo('clear');
-		if ($.fn.combobox) $('.combobox-f', target).combobox('clear');
-		if ($.fn.combotree) $('.combotree-f', target).combotree('clear');
-		if ($.fn.combogrid) $('.combogrid-f', target).combogrid('clear');
-		validate(target);
+		
+		var form = $(target);
+		var opts = $.data(target, 'form').options;
+		for(var i=opts.fieldTypes.length-1; i>=0; i--){
+			var type = opts.fieldTypes[i];
+			var field = form.find('.'+type+'-f');
+			if (field.length && field[type]){
+				field[type]('clear');
+			}
+		}
+		form.form('validate');
 	}
 	
 	function reset(target){
 		target.reset();
-		var t = $(target);
-		if ($.fn.combo){t.find('.combo-f').combo('reset');}
-		if ($.fn.combobox){t.find('.combobox-f').combobox('reset');}
-		if ($.fn.combotree){t.find('.combotree-f').combotree('reset');}
-		if ($.fn.combogrid){t.find('.combogrid-f').combogrid('reset');}
-		if ($.fn.spinner){t.find('.spinner-f').spinner('reset');}
-		if ($.fn.timespinner){t.find('.timespinner-f').timespinner('reset');}
-		if ($.fn.numberbox){t.find('.numberbox-f').numberbox('reset');}
-		if ($.fn.numberspinner){t.find('.numberspinner-f').numberspinner('reset');}
-		validate(target);
+		var form = $(target);
+		var opts = $.data(target, 'form').options;
+		for(var i=opts.fieldTypes.length-1; i>=0; i--){
+			var type = opts.fieldTypes[i];
+			var field = form.find('.'+type+'-f');
+			if (field.length && field[type]){
+				field[type]('reset');
+			}
+		}
+		form.form('validate');
 	}
 	
 	/**
@@ -227,28 +260,38 @@
 	 */
 	function setForm(target){
 		var options = $.data(target, 'form').options;
-		var form = $(target);
-		form.unbind('.form').bind('submit.form', function(){
-			setTimeout(function(){
-				ajaxSubmit(target, options);
-			}, 0);
-			return false;
+		$(target).unbind('.form');
+		if (options.ajax){
+			$(target).bind('submit.form', function(){
+				setTimeout(function(){
+					ajaxSubmit(target, options);
+				}, 0);
+				return false;
+			});
+		}
+		$(target).bind('_change.form', function(e, t){
+			options.onChange.call(this, t);
+		}).bind('change.form', function(e){
+			var t = e.target;
+			if (!$(t).hasClass('textbox-text')){
+				options.onChange.call(this, t);
+			}
 		});
+		setValidation(target, options.novalidate);
 	}
 	
-//	function validate(target){
-//		if ($.fn.validatebox){
-//			var box = $('.validatebox-text', target);
-//			if (box.length){
-//				box.validatebox('validate');
-////				box.trigger('focus');
-////				box.trigger('blur');
-//				var invalidbox = $('.validatebox-invalid:first', target).focus();
-//				return invalidbox.length == 0;
-//			}
-//		}
-//		return true;
-//	}
+	function initForm(target, options){
+		options = options || {};
+		var state = $.data(target, 'form');
+		if (state){
+			$.extend(state.options, options);
+		} else {
+			$.data(target, 'form', {
+				options: $.extend({}, $.fn.form.defaults, $.fn.form.parseOptions(target), options)
+			});
+		}
+	}
+	
 	function validate(target){
 		if ($.fn.validatebox){
 			var t = $(target);
@@ -260,26 +303,33 @@
 		return true;
 	}
 	
+	function setValidation(target, novalidate){
+		var opts = $.data(target, 'form').options;
+		opts.novalidate = novalidate;
+		$(target).find('.validatebox-text:not(:disabled)').validatebox(novalidate ? 'disableValidation' : 'enableValidation');
+	}
+	
 	$.fn.form = function(options, param){
 		if (typeof options == 'string'){
+			this.each(function(){
+				initForm(this);
+			});
 			return $.fn.form.methods[options](this, param);
 		}
 		
-		options = options || {};
 		return this.each(function(){
-			if (!$.data(this, 'form')){
-				$.data(this, 'form', {
-					options: $.extend({}, $.fn.form.defaults, options)
-				});
-			}
+			initForm(this, options);
 			setForm(this);
 		});
 	};
 	
 	$.fn.form.methods = {
+		options: function(jq){
+			return $.data(jq[0], 'form').options;
+		},
 		submit: function(jq, options){
 			return jq.each(function(){
-				ajaxSubmit(this, $.extend({}, $.fn.form.defaults, options||{}));
+				ajaxSubmit(this, options);
 			});
 		},
 		load: function(jq, data){
@@ -299,15 +349,39 @@
 		},
 		validate: function(jq){
 			return validate(jq[0]);
+		},
+		disableValidation: function(jq){
+			return jq.each(function(){
+				setValidation(this, true);
+			});
+		},
+		enableValidation: function(jq){
+			return jq.each(function(){
+				setValidation(this, false);
+			});
 		}
 	};
 	
+	$.fn.form.parseOptions = function(target){
+		var t = $(target);
+		return $.extend({}, $.parser.parseOptions(target, [{ajax:'boolean'}]), {
+			url: (t.attr('action') ? t.attr('action') : undefined)
+		});
+	};
+	
 	$.fn.form.defaults = {
+		fieldTypes: ['combobox','combotree','combogrid','datetimebox','datebox','combo',
+		        'datetimespinner','timespinner','numberspinner','spinner',
+		        'slider','searchbox','numberbox','textbox','switchbutton'],
+		novalidate: false,
+		ajax: true,
 		url: null,
+		queryParams: {},
 		onSubmit: function(param){return $(this).form('validate');},
 		success: function(data){},
 		onBeforeLoad: function(param){},
 		onLoadSuccess: function(data){},
-		onLoadError: function(){}
+		onLoadError: function(){},
+		onChange: function(target){}
 	};
 })(jQuery);
